fix(cli): extract shared translation pipeline, add preview translations, parallel locale publishing
Issue #1: Extract buildTranslatedDocsDefinition into shared module
  - Both V2 (publishDocs.ts) and ledger (publishDocsLedger.ts) now import
    from buildTranslatedDocsDefinition.ts instead of duplicating ~140 lines
    of MDX processing logic (snippet resolution, image rewrites, nav overlays)

Issue #2: Wire resolver into ledger preview path
  - publishDocsViaLedgerPreview now accepts optional resolver parameter
  - When translations are available, preview publishes them via
    finishTranslation after the base deployment completes

Issue #3: Parallel locale publishing
  - Replace sequential for...of with Promise.all in
    publishTranslationsViaLedger, matching V2's parallel approach

Issue #4: Document dual-mode double-publish
  - Add comment explaining that dual mode intentionally publishes
    translations via both V2 and ledger for migration parity
